Add redirect handlers for forum and contact
